Bytes aware deserializer
This uses the DeserializeSeed trait from serde to hold the original
input as state for the deserializer. When we encounter a &str, we know
that it was generated from the input Bytes, so we can make a new Bytes
instance by slicing it.

This will panic if the Bytes used in BytesSeed is not the same one used
for deserialization, so to avboid that issue, the BytesSeed type is not
public, and its only use enforces the input data