we should use visit_borrowed_str instead of visit_str
visit_borrowed_str is used when the data lives as long as the
deserializer, while with visit_str, it could disappear right after the
call (in the default implementation, visit_borrowed_str calls
visit_str).

apparently visit_str could be called when parsing escaped strings: they
are unescaped and written in a scratch vector, and visit_str gets a
slice of that vector. In that case instead of using a slice of the input
Bytes, we will generate a new one only for this string